requirements.txt

torch
streamlit
openai-whisper
gTTS
pydub
